[refactor] Break chain/client/src/test_utils.rs into module (#9611)
Breaking a large test_utils.rs file into multiple parts.
Additionally shifting location of `TestEnvNightshadeSetupExt` to
nearcore so that it can be used in other modules, specifically
`tools/state-viewer/src/state_dump.rs`.

Note that we can't have `TestEnvNightshadeSetupExt` in
`chain/client/src/test_utils` as it depends on NightshadeRuntime which
is a part of `nearcore` module.

`nearcore` module depends on `chain` and would create a cyclic
dependency.
