+++ integration-tests/src/tests/client/features/fix_storage_usage.rs
@@ -7,9 +7,9 @@ use near_primitives::version::ProtocolFeature;
 use near_primitives::{trie_key::TrieKey, types::AccountId};
 use near_store::{ShardUId, TrieUpdate};
 use nearcore::config::GenesisExt;
+use nearcore::test_utils::TestEnvNightshadeSetupExt;
 
 use crate::tests::client::process_blocks::set_block_protocol_version;
-use crate::tests::client::utils::TestEnvNightshadeSetupExt;
 
 fn process_blocks_with_storage_usage_fix(
     chain_id: String,
