+++ integration-tests/src/tests/client/mod.rs
@@ -15,4 +15,3 @@ mod state_dump;
 mod state_snapshot;
 mod sync_state_nodes;
 mod undo_block;
-mod utils;
