+++ integration-tests/src/tests/client/runtimes.rs
@@ -13,10 +13,9 @@ use near_primitives::network::PeerId;
 use near_primitives::test_utils::create_test_signer;
 use near_primitives::validator_signer::InMemoryValidatorSigner;
 use nearcore::config::GenesisExt;
+use nearcore::test_utils::TestEnvNightshadeSetupExt;
 use std::collections::HashMap;
 use std::sync::Arc;
-
-use crate::tests::client::utils::TestEnvNightshadeSetupExt;
 
 #[test]
 fn test_pending_approvals() {
